# docs/api/openapi.v1.yaml
openapi: 3.1.0
info:
  title: Luqa Mobile API
  version: 1.0.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: >-
    Stable native-client contract for Luqa. Authentication uses revocable
    opaque bearer tokens; the browser companion retains its Auth.js session.
servers:
  - url: /api/v1
...
paths:
...
  /money/people:
    get:
      operationId: listPeople
      tags: [Money]
      summary: List everyone, archived included, in display order
      responses:
        "200":
          description: People
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createPerson
      tags: [Money]
      summary: Add someone to split with
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePersonRequest"
      responses:
        "200":
          description: >-
            Someone with this name — or this id — already exists on the
            account; they are returned rather than duplicated, so the same
            person added offline on two devices converges on one row.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonResponse"
        "201":
          description: Person created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonResponse"
        "400":
          $ref: "#/components/responses/InvalidInput"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/IdConflict"
  /money/people/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    patch:
      operationId: updatePerson
      tags: [Money]
      summary: Rename, restyle, reorder, archive, or restore someone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePersonRequest"
      responses:
        "200":
          description: Person updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonResponse"
        "400":
          $ref: "#/components/responses/InvalidInput"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deletePerson
      tags: [Money]
      summary: Remove someone
      description: >-
        Anyone who has been on a bill is archived instead, so the history that
        produced their balance survives. `deleted` says which happened.
      responses:
        "200":
          description: Removed or archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletePersonResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /money/settlements/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    patch:
      operationId: updateSettlement
      tags: [Money]
      summary: Correct a payback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettlementRequest"
      responses:
        "200":
          description: Payback updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettlementResponse"
        "400":
          $ref: "#/components/responses/InvalidInput"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteSettlement
      tags: [Money]
      summary: Undo a payback
      responses:
        "204":
          description: Gone, or already gone.
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    mobileBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Short-lived Luqa mobile access token.
  responses:
    InvalidInput:
      description: Malformed JSON, input, or query parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Invalid, expired, revoked, or missing device credential
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: The resource does not exist, or is not owned by this user
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    IdConflict:
      description: The client-supplied id already belongs to another account
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    SettlementDirection:
      type: string
      description: Which way the money moved when a debt was paid off.
      enum: [TO_ME, FROM_ME]
    Person:
      type: object
      additionalProperties: false
      required: [id, name, color, emoji, defaultPercent, order, archived]
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: string
          pattern: "^#[0-9a-fA-F]{6}$"
        emoji:
          type: [string, "null"]
        defaultPercent:
          type: [integer, "null"]
          minimum: 0
          maximum: 100
          description: >-
            The cut of a bill this person usually carries, in whole percent.
            Null means share equally with everyone else on it.
        order:
          type: integer
        archived:
          type: boolean
...
    Settlement:
      type: object
      additionalProperties: false
      required: [id, personId, amountCents, direction, date, notes, createdAt]
      properties:
        id:
          type: string
        personId:
          type: string
        amountCents:
          type: integer
        direction:
          $ref: "#/components/schemas/SettlementDirection"
        date:
          type: string
          pattern: "^\\d{4}-\\d{2}-\\d{2}$"
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
...
    PersonResponse:
      type: object
      additionalProperties: false
      required: [person]
      properties:
        person:
          $ref: "#/components/schemas/Person"
    PersonListResponse:
      type: object
      additionalProperties: false
      required: [people]
      properties:
        people:
          type: array
          items:
            $ref: "#/components/schemas/Person"
    DeletePersonResponse:
      type: object
      additionalProperties: false
      required: [deleted]
      properties:
        deleted:
          type: boolean
          description: >-
            True when the person was removed outright, false when they were
            archived because bills or paybacks still reference them.
...
    SettlementResponse:
      type: object
      additionalProperties: false
      required: [settlement]
      properties:
        settlement:
          $ref: "#/components/schemas/Settlement"
...
    CreatePersonRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        id:
          type: string
          minLength: 8
          maxLength: 64
          pattern: "^[A-Za-z0-9_-]+$"
          description: >-
            Preferred identity for someone added offline. Honoured only when free; an existing person with the same name wins, so the response is authoritative.
        name:
          type: string
          minLength: 1
          maxLength: 60
        color:
          type: string
          pattern: "^#[0-9a-fA-F]{6}$"
        emoji:
          type: [string, "null"]
          maxLength: 8
        defaultPercent:
          type: [integer, "null"]
          minimum: 0
          maximum: 100
    UpdatePersonRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 60
        color:
          type: string
          pattern: "^#[0-9a-fA-F]{6}$"
        emoji:
          type: [string, "null"]
          maxLength: 8
        defaultPercent:
          type: [integer, "null"]
          minimum: 0
          maximum: 100
        order:
          type: integer
          minimum: 0
        archived:
          type: boolean
...
    UpdateSettlementRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        amountCents:
          type: integer
          minimum: 1
          maximum: 100000000
        direction:
          $ref: "#/components/schemas/SettlementDirection"
        date:
          type: string
          pattern: "^\\d{4}-\\d{2}-\\d{2}$"
        notes:
          type: string
          maxLength: 500
    ErrorDetail:
      type: object
      additionalProperties: true
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        issues: {}
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
...
security:
  - mobileBearer: []